requirements.txt
# Core dependencies
tokenizers==0.15.0
torch==2.2.1
transformers==4.38.2
faster-whisper==0.10.0
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
ffmpeg-python==0.2.0
python-dotenv==1.0.1

# Frontend dependencies
streamlit==1.32.0
requests==2.31.0

# Testing dependencies
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
